短信查询,短信监听,短信发送,联系人查询

+++ app/src/main/java/com/example/pinan/tempandroid/utils/DateUtil.java
@@ -15,6 +15,6 @@ public class DateUtil {
      */
     public static String dateFormat(String time) {
         SimpleDateFormat sdf = new SimpleDateFormat("yyyy-MM-dd HH-mm-ss");
-        return sdf.format(new Date(Integer.parseInt(time)));
+        return sdf.format(new Date(Long.parseLong(time)));
     }
 }
